introducing modified sessions
