--- mobile_frontend/lib/models/session.dart
@@ -1,13 +1,0 @@
-class Session{
-  int numSeanceDay, group;
-  String module, seanceType, classroom;
-  bool bySection;
-  String timeStart, timeEnd;
-
-  Session(this.numSeanceDay, this.group, this.module, this.seanceType,
-      this.classroom, this.bySection, this.timeStart, this.timeEnd);
-
-  String timeLaps(){
-    return timeStart + " - " + timeEnd;
-  }
-}
